requirements.txt
aiogram==3.3.0
matplotlib
numpy
Pillow
torch
torchvision
typing_extensions
